// _archive/jellysea/src/utils/types.ts
export enum MediaStatus {
  UNKNOWN = 1,
  PENDING = 2,
  PROCESSING = 3,
  PARTIALLY_AVAILABLE = 4,
  AVAILABLE = 5,
  DELETED = 6,
}

export enum MediaType {
  MOVIE = 'movie',
  TV = 'tv',
}

export enum MediaRequestStatus {
  PENDING = 1,
  APPROVED = 2,
  DECLINED = 3,
}

export interface DownloadingItem {
  externalId: number
  downloadId: string
  size: number
  sizeLeft: number
  status: string
  timeLeft: string
  estimatedCompletionTime: string
  episode?: {
    seasonNumber: number
    episodeNumber: number
  }
}

export interface Season {
  id: number
  seasonNumber: number
  status: MediaStatus
}

export interface User {
  id: number;
  email: string;
  username?: string;
  plexUsername?: string;
  jellyfinUsername?: string;
  displayName: string;
  avatar: string;
  permissions: number;
  userType: number;
  requestCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface MediaInfo {
  id: number;
  tmdbId: number;
  tvdbId?: number;
  mediaType: 'movie' | 'tv';
  status: MediaStatus;
  status4k?: MediaStatus;
  requests: MediaRequest[];
  plexUrl?: string;
  plexUrl4k?: string;
  iOSPlexUrl?: string;
  iOSPlexUrl4k?: string;
  serviceUrl?: string;
  serviceUrl4k?: string;
  downloadStatus?: DownloadingItem[];
  downloadStatus4k?: DownloadingItem[];
  mediaAdded?: string;
  createdAt: string;
  updatedAt: string;
  seasons?: Season[];
}

export interface MediaRequest {
  id: number;
  status: number;
  media: MediaInfo;
  type?: 'movie' | 'tv';
  seasons?: Season[];
  createdAt: string;
  updatedAt: string;
  requestedBy: User;
  is4k: boolean;
}
// ...
export interface Genre {
  id: number
  name: string
}

export interface GenreSliderItem {
  id: number
  name: string
  backdrops: string[]
}

export interface StudioResult {
  id: number
  name: string
  backdrops: string[]
}

export interface NetworkResult {
  id: number
  name: string
  backdrops: string[]
}

export interface RequestResponse {
  pageInfo: PageInfo
  results: MediaRequest[]
}

export interface ProductionCompany {
  id: number
  logoPath?: string
  name: string
  originCountry?: string
}
// ...
export interface Cast {
  id: number
  castId: number
  character: string
  creditId: string
  gender?: number
  name: string
  order: number
  profilePath?: string
}

export interface Crew {
  id: number
  creditId: string
  department: string
  gender?: number
  job: string
  name: string
  profilePath?: string
  order?: number
}

export interface Video {
  url?: string
  site: string
  key: string
  name: string
  size: number
  type: string
}

export interface Keyword {
  id: number
  name: string
}

export interface ExternalIds {
  imdbId?: string
  freebaseMid?: string
  freebaseId?: string
  tvdbId?: number
  tvrageId?: string
  facebookId?: string
  instagramId?: string
  twitterId?: string
}

export interface WatchProvider {
  displayPriority?: number
  logoPath?: string
  id: number
  name: string
}

export interface WatchProviders {
  iso_3166_1: string
  link?: string
  buy?: WatchProvider[]
  flatrate?: WatchProvider[]
}

export interface MovieDetails {
  id: number
  imdbId?: string
  adult: boolean
  backdropPath?: string
  posterPath?: string
  budget: number
  genres: Genre[]
  homepage?: string
  originalLanguage: string
  originalTitle: string
  overview?: string
  popularity: number
  relatedVideos?: Video[]
  productionCompanies: ProductionCompany[]
  productionCountries: { iso_3166_1: string; name: string }[]
  releaseDate: string
  revenue: number
  runtime?: number
  spokenLanguages: { iso_639_1: string; name: string }[]
  status: string
  tagline?: string
  title: string
  video: boolean
  voteAverage: number
  voteCount: number
  credits: {
    cast: Cast[]
    crew: Crew[]
  }
  collection?: {
    id: number
    name: string
    posterPath?: string
    backdropPath?: string
  }
  mediaInfo?: MediaInfo
  externalIds: ExternalIds
  watchProviders?: WatchProviders[]
  keywords: Keyword[]
}
// ...
export interface RTRating {
  criticsRating: string
  criticsScore: number
  audienceRating: string
  audienceScore: number
  url: string
}

export interface IMDBRating {
  criticsScore: number
  url: string
}

export interface RatingResponse {
  rt?: RTRating
  imdb?: IMDBRating
}